Rename overrides package to defaults
It's not really "overriding" values, more applying default values when
not already set.
